refactor(ui): consolidate modality + models into gen module
