--- packages/ui/src/modules/modality/index.ts
@@ -1,5 +1,0 @@
-export {
-    getModalityTheme,
-    type Modality,
-    modalityTheme,
-} from "./themes.ts";
