+++ packages/ui/tsup.config.ts
@@ -6,8 +6,7 @@ export default defineConfig({
         "auth/index": "src/modules/auth/index.ts",
         "auth/sdk": "src/modules/auth/sdk.ts",
         "app-user-menu/sdk": "src/modules/app-user-menu/sdk.ts",
-        "modality/index": "src/modules/modality/index.ts",
-        "models/index": "src/modules/models/index.ts",
+        "gen/index": "src/modules/gen/index.ts",
         "wallet/index": "src/modules/wallet/index.ts",
         "wallet/sdk": "src/modules/wallet/sdk.ts",
     },
